Add flash msg when deleating a post, fix error and reorder of imports

+++ app/views/post_view.py
@@ -1,4 +1,10 @@
-from flask import render_template, redirect, url_for, abort
+from flask import (
+                    abort,
+                    flash,
+                    render_template,
+                    redirect,
+                    url_for
+                  )
 from flask_classy import FlaskView, route
 from flask_user import login_required, current_user
 from ..models import PostModel
@@ -47,7 +53,11 @@ class Post(FlaskView):
     @login_required
     @route("/delete/<entity_id>")
     def delete(self, entity_id):
-        if int(entity_id) in current_user.posts_list:
+        if hasattr(current_user, "posts_list"
+                   ) and int(entity_id) in current_user.posts_list:
             PostModel.delete(entity_id)
-            return "Your post have been delete."
-        return "You do not have a Post with an ID {}".format(entity_id)
+            flash("Your post have been delete.", "success")
+        else:
+            flash("You do not have a Post with an ID {}".format(
+                  entity_id), "error")
+        return redirect(url_for("Main:index"))
